Ignore events and observation requests during shutdown
In particular for event sources, which are shut down when the loop itself is
being shut down, it can happen that even for correct code, events are
dispatched while the loop is shutting down. In that scenario, there is a
race that leads to an exception - either due to a rejected execution by
the executor service, or due to the sanity checks that verify correct
caller behaviour by disallowing interactions with a disposed loop.

This change adds a 'DISPOSING' state, during which events and observation
requests are ignored.