Make MessageDispatcher not forward messages when disposed.
Not all WorkRunner implementations are safe when sending in work after dispose.
Without this change, a race between MobiusLoop.dispose and dispatchEvent, where
dispatchEvent stops between checking that runState != DISPOSED but before posting
the event to the event dispatcher, and is only resumed once dispose() completes
leads to exceptions (RejectedExecutionException).