remove synchronization from DiscardAfterDisposeWrapper
The use of 'synchronized' had the effect of preventing 'dispose' and 'accept' from racing.
In some cases, this led to disposal behaviour unit tests to failing because MobiusLoop.dispose
was prevented from making progress due to an ongoing handling of an event. Removing that lock
improves test robustness and also the semantics of the wrapper.