Fix out-of-order model notifications
In a situation where the addition of an observer races with a model change, it
used to be possible for the observer to be notified of the models in the wrong
order, leaving it with the wrong data. This change fixes that issue, but
doesn't solve another less serious problem with duplicate notifications of
the same model - something that can happen normally anyway as not all
events lead to model changes.