Simplify RxJava subscribe calls, remove onError
From RxJava 2 onwards if onError is not implemented then the default behaviour is that errors are forwarded to RxJavaPlugins so it's not necessary to explicitly do that. The errors will be wrapped differently, but the cause throwable will remain the same.

There's a reason besides aesthetics for doing this: uber/RxDogTag, the RxJava stack trace improvement library, only supports observables that do not implement onError() handling. It will be easier for us to debug issues when we use the default error handling.