Make QueuingConnection non-blocking
With the previous, blocking, implementation, there was a risk of deadlock in loops using RxEventSources.
The thread calling QueuingConnection.dispose() acquires the lock in the synchronized section, the
event source locks an AtomicBoolean, and then tries to lock the QueuingConnection, while the QC tries
to acquire the AtomicBoolean lock. Making QC non-blocking should solve that.