Fix for RxConnectables not discarding emissions after dispose.
This is an issue quite similar to #146 a race condition where it was possible for Connectables
to emit values after dispose() returned. This could happen both for ObservableTransformers converted
to Connectables and vice versa. The fix is identical to the one for RxEventSources, and for good
measure the same test was added to mobius-coroutines to make sure we don't have the same issue there,
but that doesn't seem to be the case.